+++ psmm.psm1
@@ -22,6 +22,7 @@ $psmmEngineFiles = @(
     'src/Engine/Auth.ps1'
     'src/Engine/Gallery.ps1'
     'src/Engine/CloudFiles.ps1'
+    'src/Engine/SelfUpdate.ps1'
 )
 $psmmPublicFiles = @(
     'src/Public/Invoke-PSMMStartup.ps1'